server/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
psycopg2-binary==2.9.11
redis==5.0.1
python-dotenv==1.0.0
Werkzeug==3.0.1
Pillow==12.0.0
email-validator==2.1.0
